Ajout de la page de pré-jeu
