+++ nombre_mystere/lib/ui/detailscreen.dart
@@ -5,8 +5,8 @@ class DetailScreen extends StatelessWidget {
   final int id;
 
   const DetailScreen(
-      {super.key, required this.id}
-      );
+    {super.key, required this.id}
+  );
 
   @override
   Widget build(BuildContext context) {
